feat(core): fonts.rs — simple-font decoding with full AGL/encodings/Core-14 (PER-37)
- pdf/tables/: generated static tables (committed, zero build-time network):
  Adobe Glyph List (4281), base encodings (Standard/WinAnsi/MacRoman/ZapfDingbats),
  Core-14 metrics (Helvetica/Times/Symbol/ZapfDingbats families + Courier const)
- fonts.rs: Font::from_dict maps each byte code -> (Unicode, advance width) via
  base encoding + Differences + AGL + uniXXXX/uXXXX; widths from Widths array,
  else Core-14 metrics, else fallback. Type0/CID flagged is_cid (CJK -> M2)
- tools/gen_font_tables/: generator + source provenance for reproducibility
- also fix a pre-existing redundant f32 cast in objects.rs (clippy)
- 5 font unit tests; workspace clippy clean

+++ crates/pdfmuse-core/src/pdf/objects.rs
@@ -78,7 +78,7 @@ impl PdfDoc {
 fn number(o: &Object) -> Option<f32> {
     match o {
         Object::Integer(i) => Some(*i as f32),
-        Object::Real(r) => Some(*r as f32),
+        Object::Real(r) => Some(*r),
         _ => None,
     }
 }
